### 2022-07-25 #4605 @editorialbot
[PRE REVIEW]: UM-Bridge: Uncertainty quantification and modeling bridge
**Submitting author:** @linusseelinger (<a href="http://orcid.org/0000-0001-8632-8493">Linus Seelinger</a>)
**Repository:** https://github.com/UM-Bridge/umbridge
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.1
