### 2022-07-27 a329e2c
Add Jacobian evaluation to example clients and docs, mention MUQ server integration
