### 2022-07-27 #4605 @pdebuyl
Hi @kyleniemeyer . I could, but not before mid-august, and possibly after going through some backlog then. Also, this seems to be an AAS paper, what are the specifics for editing such articles?
