### 2022-07-28 1e4b663
Minor fixes in JOSS paper
