+++ joss/paper.bib
@@ -13,6 +13,7 @@
 
 
 @article{PCE,
+  doi = {10.1146/annurev.fluid.010908.165248},
   title={Uncertainty quantification and polynomial chaos techniques in computational fluid dynamics},
   author={Najm, Habib N},
   journal={Annual review of fluid mechanics},
