+++ source/inverse-benchmarks/index.rst
@@ -1,0 +1,16 @@
+.. include:: ../markdown/inverse-benchmarks.md
+   :parser: myst_parser.sphinx_
+
+For more details on the benchmarks see their individual documentation:
+
+.. toctree::
+   :maxdepth: 3
+
+   muq-beam
+   analytic-banana
+   poisson
+   analytic-funnel
+   p-poisson
+   analytic-gaussian-mixture
+   analytic-donut
+   exahype-tsunami
