+++ source/index.rst
@@ -1,0 +1,65 @@
+.. benchmarks documentation master file, created by
+   sphinx-quickstart on Wed Apr  6 13:34:13 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+UM-Bridge documentation
+======================================
+
+.. image:: logos/UM-bridge.png
+  :width: 400
+  :alt: UM-Bridge Logo
+
+.. include:: ../markdown/main.md
+   :parser: myst_parser.sphinx_
+
+Getting started
+--------------------
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+
+   /quickstart
+   /umbridge/index
+
+Propagation Benchmarks
+-----------------------
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+
+   /forward-benchmarks/index
+
+Inference Benchmarks
+-----------------------
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+
+   /inverse-benchmarks/index
+
+
+Models
+--------------------
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+
+   /models/index
+
+About
+--------------------
+
+.. toctree::
+   :glob:
+   :maxdepth: 1
+
+   /about
+
+
+
+* :ref:`search`
