+++ source/about.rst
@@ -1,0 +1,16 @@
+================
+About
+================
+
+UM-Bridge is an open-source, community project. Current core developers are listed here. We welcome contributions, please read the quickstart guide before getting started. If you want to be listed here as a contributor let us know.
+
+Core Developers
+=================
+
+- `Andy Davis <mailto:davisad@alum.mit.edu>`_
+
+- `Matthew Parno <mailto:matthew.d.parno@dartmouth.edu>`_
+
+- `Anne Reinarz <mailto:anne.k.reinarz@durham.ac.uk>`_
+
+- `Linus Seelinger <mailto:linus.seelinger@iwr.uni-heidelberg.de>`_
