### 2022-07-28 #4605 @linusseelinger
@pdebuyl The AAS reference was just a left over from the paper template we used. Sorry for the confusion!

I removed the AAS reference, and also added the missing DOI from above in bibliography.
