### 2022-08-03 #4605 @arfon
> Hi @kyleniemeyer . I could, but not before mid-august, and possibly after going through some backlog then. Also, this seems to be an AAS paper, what are the specifics for editing such articles?

Thanks @pdebuyl – I'd like to move forward and assign this to your backlog, even with the caveat about your availability. Lots of editors are in this position right now and I'd rather have this assigned to someone that leave in the waitlisted state.

@linusseelinger – heads up that even though @pdebuyl is your handling editor here, it may be another few weeks before this actually starts moving forward. Thanks for your patience!
