+++ R/NAMESPACE
@@ -1,0 +1,12 @@
+# Generated by roxygen2: do not edit by hand
+
+export(apply_hessian)
+export(apply_jacobian)
+export(evaluate)
+export(gradient)
+export(protocol_version_supported)
+export(supports_apply_hessian)
+export(supports_apply_jacobian)
+export(supports_evaluate)
+export(supports_gradient)
+importFrom(magrittr,"%>%")
