### 2022-09-05 fd255fa
Improved protocol version check for c++, Py
