### 2022-09-05 ad33b58
Fix for R test, document R client
