+++ .github/workflows/client-R.yml
@@ -1,4 +1,4 @@
-name: client-python
+name: client-R
 
 on:
   push:
@@ -24,7 +24,7 @@ jobs:
        -
         name: APT Dependencies
         run: |
-          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y R-base libcurl4-gnutls-dev libxml2-dev libfontconfig1-dev libfreetype6 libharfbuzz-dev libfribidi-dev libtiff-dev libssl-dev
+          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y r-base libcurl4-gnutls-dev libxml2-dev libfontconfig1-dev libfreetype6 libharfbuzz-dev libfribidi-dev libtiff-dev libssl-dev
        -
         name: R Dependencies
         run: |
