### 2022-09-06 e4039ef
R client: Add docs and testing, ready package for CRAN
