+++ R/R/umbridge.R
@@ -1,118 +1,207 @@
 # Need to import pipe operator
 #' @importFrom magrittr "%>%"
 
-
 umbridge_error_handler <- function(resp) {
-  print("got it")
   json <- resp %>% httr2::resp_body_json(check_type = FALSE)
   return (json$error$message)
 }
 
-# TODO: Support config!
+#' Evaluate model.
+#'
+#' @param url URL the model is running at.
+#' @param parameters Model input parameter (a list of vectors).
+#' @param config Model-specific configuration options.
+#' @return The model output (a list of vectors).
 #' @export
-evaluate <- function(url, parameters) {
+evaluate <- function(url, parameters, config = jsonlite::fromJSON("{}")) {
     stopifnot(typeof(url) == "character")
     stopifnot(typeof(parameters) == "list")
 
-    endpoint_url <- httr::modify_url(url, path="Evaluate")
-    req_json <- jsonlite::toJSON(list(input = parameters))
+    req_json <- jsonlite::toJSON(list(input = parameters, config = config))
 
-    resp_json <- httr2::request(endpoint_url) %>%
+    resp_json <- httr2::request(url) %>%
+            httr2::req_url_path_append("Evaluate") %>%
             httr2::req_body_raw(req_json) %>%
             httr2::req_error(body = umbridge_error_handler) %>%
             httr2::req_perform() %>% httr2::resp_body_json(check_type = FALSE)
     return(resp_json$output)
 }
 
-# TODO: Support config!
+#' Evaluate gradient of target functional depending on model.
+#'
+#' @param url URL the model is running at.
+#' @param out_wrt Output variable to take gradient with respect to.
+#' @param in_wrt Input variable to take gradient with respect to.
+#' @param parameters Model input parameter (a list of vectors).
+#' @param sens Sensitivity of target functional with respect to model output.
+#' @param config Model-specific configuration options.
+#' @return Gradient of target functional.
 #' @export
-gradient <- function(url, out_wrt, in_wrt, parameters, sens) {
+gradient <- function(url, out_wrt, in_wrt, parameters, sens, config = jsonlite::fromJSON("{}")) {
     stopifnot(typeof(url) == "character")
     stopifnot(typeof(parameters) == "list")
 
-    endpoint_url <- httr::modify_url(url, path="Gradient")
-    # Have to  jsonlite::unbox in order to interpret values as scalars and not lists
-    req_json <- jsonlite::toJSON(list(outWrt =  jsonlite::unbox(out_wrt), inWrt =  jsonlite::unbox(in_wrt), input = parameters, sens = sens))
-    resp_json <- httr2::request(endpoint_url) %>%
+    # Have to jsonlite::unbox in order to interpret values as scalars and not lists
+    req_json <- jsonlite::toJSON(list(outWrt = jsonlite::unbox(out_wrt), inWrt = jsonlite::unbox(in_wrt), input = parameters, sens = sens))
+    resp_json <- httr2::request(url) %>%
+            httr2::req_url_path_append("Gradient") %>%
             httr2::req_body_raw(req_json) %>%
             httr2::req_error(body = umbridge_error_handler) %>%
             httr2::req_perform() %>% httr2::resp_body_json(check_type = FALSE)
     return(resp_json$output)
 }
 
-# TODO: Support config!
+#' Evaluate Jacobian of model.
+#'
+#' @param url URL the model is running at.
+#' @param out_wrt Output variable to take Jacobian with respect to.
+#' @param in_wrt Input variable to take Jacobian with respect to.
+#' @param parameters Model input parameter (a list of vectors).
+#' @param vec Vector to multiply Jacobian by.
[…]
