+++ R/DESCRIPTION
@@ -1,15 +1,19 @@
 Package: umbridge
-Title: R integration for the UM-Bridge protocol
+Title: R Integration for the UM-Bridge Protocol
 Version: 0.9.0.0000
 Authors@R:
     person("Linus", "Seelinger", , "mail@linusseelinger.de", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-8632-8493"))
-Description: This package provides a convenient wrapper for the UM-Bridge protocol. UM-Bridge is a protocol designed for coupling uncertainty quantification (or statistical / optimization) software to numerical models. A model is represented as a mathematical function with optional support for derivatives via Jacobian actions etc.
-License: `use_mit_license()`
+Description: A convenient wrapper for the UM-Bridge protocol. UM-Bridge is a protocol designed for coupling uncertainty quantification (or statistical / optimization) software to numerical models. A model is represented as a mathematical function with optional support for derivatives via Jacobian actions etc.
+License: MIT + file LICENSE
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.1
+BugReports: https://www.github.com/um-bridge
 Imports:
-    dplyr,
 	httr2,
-	jsonlite
+	jsonlite,
+    magrittr
+Suggests: 
+    testthat (>= 3.0.0)
+Config/testthat/edition: 3
