+++ R/NAMESPACE
@@ -4,6 +4,8 @@ export(apply_hessian)
 export(apply_jacobian)
 export(evaluate)
 export(gradient)
+export(model_input_sizes)
+export(model_output_sizes)
 export(protocol_version_supported)
 export(supports_apply_hessian)
 export(supports_apply_jacobian)
