### 2022-09-06 72feb77
Run R package test in CI
