### 2022-09-06 49c9f9e
R: Add generated docs and tests
