+++ R/man/supports_apply_hessian.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{supports_apply_hessian}
+\alias{supports_apply_hessian}
+\title{Check if model supports Hessian action.}
+\usage{
+supports_apply_hessian(url)
+}
+\arguments{
+\item{url}{URL the model is running at.}
+}
+\value{
+TRUE if model supports Hessian action, FALSE otherwise.
+}
+\description{
+Check if model supports Hessian action.
+}
