+++ R/man/evaluate.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{evaluate}
+\alias{evaluate}
+\title{Evaluate model.}
+\usage{
+evaluate(url, parameters, config = jsonlite::fromJSON("{}"))
+}
+\arguments{
+\item{url}{URL the model is running at.}
+
+\item{parameters}{Model input parameter (a list of vectors).}
+
+\item{config}{Model-specific configuration options.}
+}
+\value{
+The model output (a list of vectors).
+}
+\description{
+Evaluate model.
+}
