+++ R/man/apply_hessian.Rd
@@ -1,0 +1,40 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{apply_hessian}
+\alias{apply_hessian}
+\title{Evaluate Hessian of model.}
+\usage{
+apply_hessian(
+  url,
+  out_wrt,
+  in_wrt1,
+  in_wrt2,
+  parameters,
+  sens,
+  vec,
+  config = jsonlite::fromJSON("{}")
+)
+}
+\arguments{
+\item{url}{URL the model is running at.}
+
+\item{out_wrt}{Output variable to take Hessian with respect to.}
+
+\item{in_wrt1}{First input variable to take Hessian with respect to.}
+
+\item{in_wrt2}{Second input variable to take Hessian with respect to.}
+
+\item{parameters}{Model input parameter (a list of vectors).}
+
+\item{sens}{Sensitivity with respect to output.}
+
+\item{vec}{Vector to multiply Hessian by.}
+
+\item{config}{Model-specific configuration options.}
+}
+\value{
+Hessian with respect to given inputs and outputs, applied to given sensitivity and vector.
+}
+\description{
+Evaluate Hessian of model.
+}
