+++ R/man/supports_evaluate.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{supports_evaluate}
+\alias{supports_evaluate}
+\title{Check if model supports evaluation.}
+\usage{
+supports_evaluate(url)
+}
+\arguments{
+\item{url}{URL the model is running at.}
+}
+\value{
+TRUE if model supports evaluation, FALSE otherwise.
+}
+\description{
+Check if model supports evaluation.
+}
