+++ R/man/model_input_sizes.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{model_input_sizes}
+\alias{model_input_sizes}
+\title{Retrieve model's input dimensions.}
+\usage{
+model_input_sizes(url)
+}
+\arguments{
+\item{url}{URL the model is running at.}
+}
+\value{
+List of input dimensions.
+}
+\description{
+Retrieve model's input dimensions.
+}
