+++ R/man/supports_apply_jacobian.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/umbridge.R
+\name{supports_apply_jacobian}
+\alias{supports_apply_jacobian}
+\title{Check if model supports Jacobian action.}
+\usage{
+supports_apply_jacobian(url)
+}
+\arguments{
+\item{url}{URL the model is running at.}
+}
+\value{
+TRUE if model supports Jacobian action, FALSE otherwise.
+}
+\description{
+Check if model supports Jacobian action.
+}
