### 2022-09-06 c5cdf6e
R: Update pkg. description
