+++ R/DESCRIPTION
@@ -1,6 +1,7 @@
 Package: umbridge
 Title: R Integration for the UM-Bridge Protocol
 Version: 0.9.0.0000
+Maintainer: Linus Seelinger <mail@linusseelinger.de>
 Authors@R:
     person("Linus", "Seelinger", , "mail@linusseelinger.de", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-8632-8493"))
@@ -9,11 +10,11 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.1
-BugReports: https://www.github.com/um-bridge
+BugReports: https://github.com/um-bridge
 Imports:
 	httr2,
 	jsonlite,
     magrittr
-Suggests: 
+Suggests:
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
