+++ .github/workflows/client-R.yml
@@ -30,13 +30,13 @@ jobs:
         run: |
           R -e 'install.packages("devtools")'
        -
-        name: Test R package
-        run: |
-          cd R && R -e 'devtools::test()'
-       -
         name: Build and install R package
         run: |
           cd R && R -e 'devtools::install()'
+       -
+        name: Test R package
+        run: |
+          cd R && R -e 'devtools::test()'
        -
         name: Run client
         run: |
