### 2022-09-07 20c1219
Fix gradient in testmodel-python
