+++ .github/workflows/client-python-pymc.yml
@@ -1,0 +1,31 @@
+name: client-python
+
+on:
+  push:
+    branches:
+      - 'main'
+
+jobs:
+
+  test:
+    runs-on: ubuntu-latest
+    container: ubuntu:latest
+
+    services:
+      model:
+        image: linusseelinger/benchmark-analytic-funnel:latest
+        ports:
+        - 4243:4243
+
+    steps:
+       -
+        name: Checkout
+        uses: actions/checkout@v2
+       -
+        name: Dependencies
+        run: |
+          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install umbridge[pymc]
+       -
+        name: Build and run
+        run: |
+          cd clients/python && python3 pymc-client.py http://model:4243
