### 2022-09-09 5fa7015
Add new integrations to JOSS paper
