### 2022-09-09 5ba4445
Some more JOSS paper improvements, add PyMC deps to test
