+++ joss/paper.bib
@@ -86,4 +86,13 @@ doi = {10.4208/cicp.2009.v6.p826}
   number={239},
   pages={2},
   year={2014}
-} 
+}
+
+@article{PyMC,
+author = {Salvatier, John and Wiecki, Thomas and Fonnesbeck, Christopher},
+year = {2016},
+month = {01},
+pages = {},
+title = {Probabilistic programming in Python using PyMC3},
+doi = {10.7287/PEERJ.PREPRINTS.1686V1}
+}
