### 2022-09-09 ccc7d4f
JOSS: Update citation
