### 2022-09-09 #4605 @linusseelinger
(some minor updates to the paper)
