### 2022-09-09 #4748 @editorialbot
[REVIEW]: UM-Bridge: Uncertainty quantification and modeling bridge
**Submitting author:** @linusseelinger (<a href="http://orcid.org/0000-0001-8632-8493">Linus Seelinger</a>)
**Repository:** https://github.com/UM-Bridge/umbridge
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.2
