### 2022-09-09 #4748 @pdebuyl
Hi @georgiastuart , @Himscipy ,

During the review, feel free to ask questions here to the authors directly if you need clarifications, or to me regarding the editorial part. If you open issues on the code's repository, please report them and/or their resolution here as well so that I can follow that part as well.

Thanks for reviewing for JOSS!
