### 2022-09-13 234047a
Update R package as submitted to CRAN
