+++ clients/R/client.R
@@ -12,18 +12,18 @@ url <- args[1]
 stopifnot(protocol_version_supported(url))
 
 # Print what features the model supports
-print("Evaluation support")
+message("Evaluation support")
 supports_evaluate(url)
-print("Gradient support")
+message("Gradient support")
 supports_gradient(url)
-print("Jacobian support")
+message("Jacobian support")
 supports_apply_jacobian(url)
-print("Hessian support")
+message("Hessian support")
 supports_apply_hessian(url)
 
-print("Input dimensions")
+message("Input dimensions")
 model_input_sizes(url)
-print("Output dimensions")
+message("Output dimensions")
 model_output_sizes(url)
 
 # Define a parameter
@@ -31,15 +31,18 @@ param <- list()
 param[[1]] <- c(100.0, 18.0)
 
 # Evaluate model for parameter
+message("Evaluating model")
 output <- evaluate(url, param)
 print(output[[1]][[1]])
 
+message("Evaluating model with config")
 config = list(scale = jsonlite::unbox(2.0))
 output <- evaluate(url, param, config)
 print(output[[1]][[1]])
 
 # If model supports Jacobian actions, apply Jacobian at param to a vector
 if (supports_apply_jacobian(url)) {
+  message("Jacobian action")
   output <- apply_jacobian(url, 0, 0, param, c(1.0, 4.0))
   print(output[[1]][[1]])
 }
