### 2022-09-13 556dfbb
Add test for CRAN package
