### 2022-09-13 0b4e53d
Add CRAN installation to readme
