### 2022-09-20 05bbe96
Update to protocol ver. 1.0, supporting multiple models per server and config dependent dimensions
