+++ lib/umbridge.h
@@ -18,18 +18,18 @@ namespace umbridge {
 
   class Model {
   public:
+    Model(std::string name) : name(name) {}
 
-    Model(const std::vector<std::size_t> inputSizes, const std::vector<std::size_t> outputSizes)
-    : inputSizes(inputSizes), outputSizes(outputSizes)
-    {}
+    virtual std::vector<std::size_t> GetInputSizes(const json& config_json = json()) const = 0;
+    virtual std::vector<std::size_t> GetOutputSizes(const json& config_json = json()) const = 0;
 
-    virtual void Evaluate(const std::vector<std::vector<double>>& inputs,
+    virtual std::vector<std::vector<double>> Evaluate(const std::vector<std::vector<double>>& inputs,
                           json config_json = json()) {
       (void)inputs; (void)config_json; // Avoid unused argument warnings
       throw std::runtime_error("Evaluate was called, but not implemented by model!");
     }
 
-    virtual void Gradient(unsigned int outWrt,
+    virtual std::vector<double> Gradient(unsigned int outWrt,
                           unsigned int inWrt,
                           const std::vector<std::vector<double>>& inputs,
                           const std::vector<double>& sens,
@@ -38,7 +38,7 @@ namespace umbridge {
       throw std::runtime_error("Gradient was called, but not implemented by model!");
     }
 
-    virtual void ApplyJacobian(unsigned int outWrt,
+    virtual std::vector<double> ApplyJacobian(unsigned int outWrt,
                               unsigned int inWrt,
                               const std::vector<std::vector<double>>& inputs,
                               const std::vector<double>& vec,
@@ -47,7 +47,7 @@ namespace umbridge {
       throw std::runtime_error("ApplyJacobian was called, but not implemented by model!");
     }
 
-    virtual void ApplyHessian(unsigned int outWrt,
+    virtual std::vector<double> ApplyHessian(unsigned int outWrt,
                               unsigned int inWrt1,
                               unsigned int inWrt2,
                               const std::vector<std::vector<double>>& inputs,
@@ -63,30 +63,93 @@ namespace umbridge {
     virtual bool SupportsApplyJacobian() {return false;}
     virtual bool SupportsApplyHessian() {return false;}
 
-    const std::vector<std::size_t> inputSizes;
-    const std::vector<std::size_t> outputSizes;
+    std::string GetName() const {return name;}
 
-    std::vector<std::vector<double>> outputs;
-    std::vector<double> gradient;
-    std::vector<double> jacobianAction;
-    std::vector<double> hessAction;
+  protected:
+    std::string name;
   };
 
   // Client-side Model connecting to a server for the actual evaluations etc.
   class HTTPModel : public Model {
   public:
 
-    HTTPModel(std::string host, httplib::Headers headers = httplib::Headers())
-    : Model(read_input_size(host, headers), read_output_size(host, headers)), host(host), headers(headers)
+    HTTPModel(std::string host, std::string name, httplib::Headers headers = httplib::Headers())
+    : Model(name), cli(host.c_str()), headers(headers)
     {
-      outputs.resize(outputSizes.size());
-      retrieveSupportedFeatures();
-    }
+      if (auto res = cli.Get("/Info", headers)) {
+        json response = json::parse(res->body);
 
-    void Evaluate(const std::vector<std::vector<double>>& inputs, json config_json = json()) override {
-      httplib::Client cli(host.c_str());
+        if (response.value<double>("protocolVersion",0) != 1.0)
+          throw std::runtime_error("Model protocol version not supported!");
+
+        // Check if requested model is available on server
+        std::vector<std::string> models = response["models"];
+        if (std::find(models.begin(), models.end(), name) == models.end()) {
+          std::string model_names = "";
[…]
