+++ R/R/umbridge.R
@@ -9,15 +9,17 @@ umbridge_error_handler <- function(resp) {
 #' Evaluate model.
 #'
 #' @param url URL the model is running at.
+#' @param name Name of the desired model.
 #' @param parameters Model input parameter (a list of vectors).
 #' @param config Model-specific configuration options.
 #' @return The model output (a list of vectors).
 #' @export
-evaluate <- function(url, parameters, config = jsonlite::fromJSON("{}")) {
+evaluate <- function(url, name, parameters, config = jsonlite::fromJSON("{}")) {
     stopifnot(typeof(url) == "character")
+    stopifnot(typeof(name) == "character")
     stopifnot(typeof(parameters) == "list")
 
-    req_json <- jsonlite::toJSON(list(input = parameters, config = config))
+    req_json <- jsonlite::toJSON(list(name = name, input = parameters, config = config))
 
     resp_json <- httr2::request(url) %>%
             httr2::req_url_path_append("Evaluate") %>%
@@ -30,6 +32,7 @@ evaluate <- function(url, parameters, config = jsonlite::fromJSON("{}")) {
 #' Evaluate gradient of target functional depending on model.
 #'
 #' @param url URL the model is running at.
+#' @param name Name of the desired model.
 #' @param out_wrt Output variable to take gradient with respect to.
 #' @param in_wrt Input variable to take gradient with respect to.
 #' @param parameters Model input parameter (a list of vectors).
@@ -37,12 +40,13 @@ evaluate <- function(url, parameters, config = jsonlite::fromJSON("{}")) {
 #' @param config Model-specific configuration options.
 #' @return Gradient of target functional.
 #' @export
-gradient <- function(url, out_wrt, in_wrt, parameters, sens, config = jsonlite::fromJSON("{}")) {
+gradient <- function(url, name, out_wrt, in_wrt, parameters, sens, config = jsonlite::fromJSON("{}")) {
     stopifnot(typeof(url) == "character")
+    stopifnot(typeof(name) == "character")
     stopifnot(typeof(parameters) == "list")
 
     # Have to jsonlite::unbox in order to interpret values as scalars and not lists
-    req_json <- jsonlite::toJSON(list(outWrt = jsonlite::unbox(out_wrt), inWrt = jsonlite::unbox(in_wrt), input = parameters, sens = sens))
+    req_json <- jsonlite::toJSON(list(name = name, outWrt = jsonlite::unbox(out_wrt), inWrt = jsonlite::unbox(in_wrt), input = parameters, sens = sens, config = config))
     resp_json <- httr2::request(url) %>%
             httr2::req_url_path_append("Gradient") %>%
             httr2::req_body_raw(req_json) %>%
@@ -54,6 +58,7 @@ gradient <- function(url, out_wrt, in_wrt, parameters, sens, config = jsonlite::
 #' Evaluate Jacobian of model.
 #'
 #' @param url URL the model is running at.
+#' @param name Name of the desired model.
 #' @param out_wrt Output variable to take Jacobian with respect to.
 #' @param in_wrt Input variable to take Jacobian with respect to.
 #' @param parameters Model input parameter (a list of vectors).
@@ -61,12 +66,13 @@ gradient <- function(url, out_wrt, in_wrt, parameters, sens, config = jsonlite::
 #' @param config Model-specific configuration options.
 #' @return Jacobian with respect to given input and output variables, applied to given vector.
 #' @export
-apply_jacobian <- function(url, out_wrt, in_wrt, parameters, vec, config = jsonlite::fromJSON("{}")) {
+apply_jacobian <- function(url, name, out_wrt, in_wrt, parameters, vec, config = jsonlite::fromJSON("{}")) {
     stopifnot(typeof(url) == "character")
[…]

[…]
