### 2022-09-21 50e5df0
Python: Better type checking for input / output parameters
