### 2022-09-22 2ca4a11
Minor improvement to c++ interface
