### 2022-09-22 b3d7ab9
Fix protocol conformity image tags
