### 2022-09-22 ed15cec
Fix c++ client
