### 2022-09-22 5bdae78
Fix handling of model name in R
