### 2022-09-25 3600976
Unify test models/benchmark a bit, pass through forward model in benchmark server
