+++ benchmarks/testbenchmark/benchmark-server.py
@@ -13,7 +13,7 @@ class Benchmark(umbridge.Model):
         return [1]
 
     def __call__(self, parameters, config):
-        model = umbridge.HTTPModel(self.model_url, "TestModel")
+        model = umbridge.HTTPModel(self.model_url, "forward")
         posterior = scipy.stats.norm.logpdf(model(parameters)[0][0], 2.0, 1)  # logpdf args: x, loc, scale
         return [[posterior]]
 
@@ -22,4 +22,10 @@ class Benchmark(umbridge.Model):
 
 benckmark = Benchmark("http://localhost:4242")
 
-umbridge.serve_models([benckmark], 4243)
+# For convenience, we can expose the forward model as part of the benchmark server in addition to the posterior.
+# This can be achieved by connecting to the forward model server as usual,
+# and then simply adding the forward model to the list of models to be served by the benchmark server.
+# Calls to the forward model will then be passed through to the forward model server.
+forward_model = umbridge.HTTPModel("http://localhost:4242", "forward")
+
+umbridge.serve_models([benckmark,forward_model], 4243)
