### 2022-09-25 e6e2c31
Fix test input size
