### 2022-09-25 dd5dc6e
Allow querying available models in Python and c++
