### 2022-09-30 ec44202
Update curl commands
