### 2022-09-30 1e2b5dc
Fix model name check in python client
