### 2022-10-25 253c92d
Add (single) thread for model execution in parallel to basic HTTP handling
