### 2022-10-27 10c5237
C++: Always send at least empty config, improved error handling
