+++ lib/umbridge.h
@@ -20,11 +20,11 @@ namespace umbridge {
   public:
     Model(std::string name) : name(name) {}
 
-    virtual std::vector<std::size_t> GetInputSizes(const json& config_json = json()) const = 0;
-    virtual std::vector<std::size_t> GetOutputSizes(const json& config_json = json()) const = 0;
+    virtual std::vector<std::size_t> GetInputSizes(const json& config_json = json::parse("{}")) const = 0;
+    virtual std::vector<std::size_t> GetOutputSizes(const json& config_json = json::parse("{}")) const = 0;
 
     virtual std::vector<std::vector<double>> Evaluate(const std::vector<std::vector<double>>& inputs,
-                          json config_json = json()) {
+                          json config_json = json::parse("{}")) {
       (void)inputs; (void)config_json; // Avoid unused argument warnings
       throw std::runtime_error("Evaluate was called, but not implemented by model!");
     }
@@ -33,7 +33,7 @@ namespace umbridge {
                           unsigned int inWrt,
                           const std::vector<std::vector<double>>& inputs,
                           const std::vector<double>& sens,
-                          json config_json = json()) {
+                          json config_json = json::parse("{}")) {
       (void)outWrt; (void)inWrt; (void)inputs; (void)sens; (void)config_json; // Avoid unused argument warnings
       throw std::runtime_error("Gradient was called, but not implemented by model!");
     }
@@ -42,7 +42,7 @@ namespace umbridge {
                               unsigned int inWrt,
                               const std::vector<std::vector<double>>& inputs,
                               const std::vector<double>& vec,
-                              json config_json = json()) {
+                              json config_json = json::parse("{}")) {
       (void)outWrt; (void)inWrt; (void)inputs; (void)vec; (void)config_json; // Avoid unused argument warnings
       throw std::runtime_error("ApplyJacobian was called, but not implemented by model!");
     }
@@ -53,7 +53,7 @@ namespace umbridge {
                               const std::vector<std::vector<double>>& inputs,
                               const std::vector<double>& sens,
                               const std::vector<double>& vec,
-                              json config_json = json()) {
+                              json config_json = json::parse("{}")) {
       (void)outWrt; (void)inWrt1; (void)inWrt2; (void)inputs; (void)sens; (void)vec; (void)config_json; // Avoid unused argument warnings
       throw std::runtime_error("ApplyHessian was called, but not implemented by model!");
     }
@@ -91,21 +91,14 @@ namespace umbridge {
     HTTPModel(std::string host, std::string name, httplib::Headers headers = httplib::Headers())
     : Model(name), cli(host.c_str()), headers(headers)
     {
-      if (auto res = cli.Get("/Info", headers)) {
-        json response = json::parse(res->body);
-
-        // Check if requested model is available on server
-        std::vector<std::string> models = SupportedModels(host, headers);
-        if (std::find(models.begin(), models.end(), name) == models.end()) {
-          std::string model_names = "";
-          for (auto& m : models) {
-            model_names += "'" + m + "' ";
-          }
-          throw std::runtime_error("Model " + name + " not found on server! Available models: " + model_names + ".");
+      // Check if requested model is available on server
+      std::vector<std::string> models = SupportedModels(host, headers);
+      if (std::find(models.begin(), models.end(), name) == models.end()) {
+        std::string model_names = "";
+        for (auto& m : models) {
+          model_names += "'" + m + "' ";
         }
-
-      } else {
-        throw std::runtime_error("GET Info failed with error type '" + to_string(res.error()) + "'");
+        throw std::runtime_error("Model " + name + " not found on server! Available models: " + model_names + ".");
       }
 
[…]
