### 2022-10-27 #4748 @georgiastuart
@linusseelinger, you can disregard the above, I found Matt's contributions in the benchmarks repo.

A few questions / comments...

1. I'm curious whether it would be simple to implement a straightforward, password-based authentication to the server?
2. You have several good examples, but it took a lot of clicking around to find the source code for the examples. Please link from the documentation to the appropriate directory in the benchmarks repo, or (even better) link and inline the files into the documentation. I think that's more useful than just having a mathematical explanation and output pictures.
3. Can you add a discussion about the format expected for the `parameters` and `config` function arguments and the return values for each method in the server? Apologies if I missed this in the docs somewhere.

I intend to look at this again and try to set up an example with my own python-based model, but the above are my first impressions. Very cool work!
