### 2022-10-27 #4748 @georgiastuart
Another question, is UM-Bridge capable of using mpi-parallelized models?
