+++ CONTRIBUTING.md
@@ -1,0 +1,38 @@
+# Contributing to UM-Bridge
+
+The goal of UM-Bridge is to connect software packages, research projects and communities. Contributions are therefore more than welcome!
+
+We are happy to help with any questions regarding contributions or usage of UM-Bridge.
+
+This also holds for the closely related benchmark library, whose sources can be found in the [benchmarks repository](https://github.com/UM-Bridge/benchmarks).
+
+Possible contributions might include:
+* Fixes and extensions
+* Support for new languages or specific software packages
+* New benchmarks and models for the UM-Bridge benchmark library
+
+## Contact
+
+You can reach the UM-Bridge team via
+
+* [slack channel linked on the docs main page](https://um-bridge-benchmarks.readthedocs.io/en/docs/)
+* [issue tracker on GitHub](https://github.com/UM-Bridge/umbridge/issues)
+* [direct email to the developers](https://um-bridge-benchmarks.readthedocs.io/en/docs/about.html)
+
+## Documentaiton
+
+The [documentation](https://um-bridge-benchmarks.readthedocs.io/en/docs/) covers both UM-Bridge itself as well as the benchmark library.
+
+It is built from markdown files in both repositories.
+
+## Contributions
+
+Contributions may be submitted via [pull requests on GitHub](https://github.com/UM-Bridge/umbridge/pulls).
+
+Testing (and deployment, when appropriate) is conducted via GitHub workflows.
+
+## Bug reports and feature requests
+
+Please send bug reports and feature requests via [issue tracker on GitHub](https://github.com/UM-Bridge/umbridge/issues).
+
+If possible, provide instructions for how to reproduce an issue. For issues that occur with specific models, it is helpful to refer to a container from the benchmark library that exposes the issue.
