### 2022-10-30 #4748 @linusseelinger
@georgiastuart Thanks a lot for your comments!

1. Authentication has not been our focus so far, we are largely working in fully trusted environments. However, I have been using authentication for running UM-Bridge model containers in a cloud environment. Here, the load balancer I'm using can check for a password/access token in HTTP headers before passing requests through to the models themselves. I believe that's the best solution, since that way the security aspect is handled by a specialized component instead of UM-Bridge itself. Do you have a specific application in mind?
2. Good point, we'll update the docs accordingly!
3. Those are (somewhat) mentioned in the clients docs, but I'll make the docs more consistent and add more detail to the model side.
4. MPI models with UM-Bridge are certainly possible, we currently have two approaches in use: Either the UM-Bridge server is a thin wrapper that calls a separate MPI parallel code (this may all happen inside a container, so is transparent to the user); or UM-Bridge runs on rank zero, distributing incoming parameters via MPI broadcast. Regarding containers: MPI inside a single container just works. MPI across multiple containers (as needed for multiple hardware nodes) is not trivial, but fully supported by our kubernetes setup (you have to inherit from a specific choice of base images in that case. Beyond that, our kubernetes config takes care of everything. You can just plug in your custom image, and your code will see all MPI ranks across nodes/containers as usual).
