### 2022-11-02 0e6e055
Print supported models in python client; needs recent python package
