+++ .github/workflows/pip-update.yml
@@ -34,6 +34,7 @@ jobs:
     - name: Build package
       run: python -m build
     - name: Publish package
+      if: startsWith(github.ref, 'refs/tags')
       uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
       with:
         user: __token__
