### 2022-11-04 #4748 @Himscipy
Hi @linusseelinger

Here few comments related to the software paper and repository;

- Could you mention how the UM-Bridge compares with the existing state of the field tools for UQ in the paper? This will be useful in addressing one of the requirements (State of the field: Do the authors describe..) in the paper.

- The repository needs unit tests /automated tests for code checking. I would recommend adding a unit test and also the code coverage as a software testing metric associated with the software repo.  This will be useful for debugging and future development of the software.
