### 2022-11-14 4bee154
Update kubernetes setup
