### 2022-11-14 d1b8c6b
Temporarily remove authentication from ingress service
