### 2022-11-14 5ca2ef2
Remove obsolete yaml
