### 2022-11-16 fcfa03c
extend input_value to vector
