### 2022-11-16 97be205
Allow vector input to test_model_schema
