### 2022-11-16 b10e658
PyMC: Run with 1 core, more reasonable for basic example
