### 2022-11-19 d9b85fb
Clean up k8s config
