+++ kubernetes/launch-exahype-tsunami-nfs.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-for i in $(seq 1 3)
-do cat exahype-tsunami-nfs.yaml | SHARED_FILE_PREFIX=$i envsubst | kubectl apply -f - &
-done
-
-wait
