+++ kubernetes/multinode-mpi-model.yaml
@@ -12,6 +12,9 @@ spec:
     Launcher:
       replicas: 1
       template:
+        metadata:
+          labels:
+            app: model
         spec:
           volumes:
           - name: my-pvc-nfs
