### 2022-11-19 2e8ffa2
Switch qmcpy from repo to pypi version
