### 2022-11-21 a9b95f3
Clean up kubernetes docs
