+++ kubernetes/shutdown-multinode-mpi-model.sh
@@ -1,2 +1,0 @@
-#/bin/sh
-kubectl delete MPIJob --all
