+++ kubernetes/README.md
@@ -1,119 +1,134 @@
-# Running on GKE
+# Cloud HPC
 
-These instructions are intended for users who have not used GKE before. If you are familiar with GKE you will not need most of these instructions, instead just have a look at the `setup.sh` bash script which contains our commonly used commands.
+These instructions show how to run the UM-Bridge kubernetes configuration. It assumes that you have a kubernetes cluster available, for example by following our insructions for setting up Google Kubernetes Engine (GKE).
 
-## Step 1: Install cloud shell
+## Step 1: Clone UM-Bridge
 
-To start you will need access to cloud shell. On ubuntu this can be snap installed:
-
-``sudo snap install google-cloud-cli --classic``
-
-It is also available to be downloaded directly from Google.
-
-
-## Step 2: Initialise and run cloud shell
-
-Next we need to initialise and set up our session in cloud shell. To initialise run:
-
-``gcloud init``
-
-It will ask you to log into your Google account and set the billing account. If you don't have one already you will need to create one.
-
-It will then ask you to set up a default location. We generally use `europe-west2-c`, but this depends on your location.
-
-Next you start an interactive session:
-
-``gcloud cloud-shell ssh --authorize-session``
-
-In parallel, it can be convenient to see whats happening in the [GUI](https://console.cloud.google.com/kubernetes/).
-
-## Step 3: Initialise a cluster
-
-Next we want to initialise a cluster to run our umbridge instances on. To start we will clone umbridge by running:
+First, clone the UM-Bridge repository by running:
 
 ``git clone https://github.com/UM-Bridge/umbridge.git``
 
-In the umbridge project, the folder `kubernetes` contains everything that is needed to set up.
+In the cloned repository, navigate to the folder `kubernetes`. It contains all that is needed in the following.
 
-- If there are running clusters you may want to delete them before setting up a new one.
+## Step 2: Set up load balancer
 
-The bash script `setup.sh` contains most commands you will need. Before continuing have a look into the scirpt. If you want to use an existing cluster proceed to step 3b.
+First, retrieve HAProxy:
 
-### Step 3a: Setting up from scratch
+``
+helm repo add haproxytech https://haproxytech.github.io/helm-charts
+helm repo update
 
-To create a cluster we first need to figure out what we want. By default the setup.sh has chosen to create a cluster with the name `testcluster` with machine-type `c2-standard-4`. This is a 4 core machine. For more options in regards to machine types check [here](https://cloud.google.com/compute/docs/compute-optimized-machines). To change the number of nodes choose `num-nodes`.
+helm install kubernetes-ingress haproxytech/kubernetes-ingress \
+--create-namespace \
+--namespace haproxy-controller \
+--set controller.service.type=LoadBalancer \
+--set controller.replicaCount=1 \
+--set defaultBackend.replicaCount=1 \
+--set controller.logging.level=debug \
+--set controller.ingressClass=haproxy
+``
 
-Once you have made the changes you want run `create_mpi_cluster` (either by commenting in the line in setup.sh or by copying out the corresponding command).
+Then, start HAProxy with the configuration provided by UM-Bridge:
 
-Next enable MPI by running `deploy_mpi_operator`.
+``
+kubectl apply -f ./setup/svc-ingress.yml
+``
+
+## Step 3: Run model instances
+
+To start model instances, run:
+
+``
+kubectl create -f model.yaml
+``
+
+You can check the status of the model instances by running:
+
+``
+kubectl get pods
+``
 
 
-## Step 3b (only if not setting up from scratch): Use an existing Cluster
+They can be again via:
 
-If the cluster is currently sized down to reduce resources, resize it:
+``
[…]
