### 2022-11-21 6f8bf12
Remove now redundant setup.sh
