### 2022-11-21 1d8b097
Minor cleanup
