### 2022-11-21 3c8a710
Docs: Nicer bash commands in k8s setup
