### 2022-11-21 fba373b
Improved client/server docs
