### 2022-11-21 #4748 @linusseelinger
@georgiastuart Examples now link to the respective source code (turns out pulling source files directly from the repo is not trivial in our docs system, so we'll go with plan B for now). Types are now better documented, in particular for Python client and server where it's not obvious to see what is expected. The Python code also enforces types by explicitly checking, so I hope that's sufficient to guide new users. I hope that covers your comments about docs / types?
