### 2022-11-21 ce98bd4
Add full client testing for Py and C++
