+++ .github/workflows/client-python-example.yml
@@ -1,4 +1,4 @@
-name: client-python
+name: client-python-example
 
 on:
   push:
@@ -13,7 +13,7 @@ jobs:
 
     services:
       model:
-        image: linusseelinger/model-muq-beam:latest
+        image: linusseelinger/model-exahype-tsunami:latest
         ports:
         - 4242:4242
 
@@ -24,8 +24,8 @@ jobs:
        -
         name: Dependencies
         run: |
-          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install umbridge pytest
+          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install umbridge
        -
         name: Build and run
         run: |
-          cd testing/clients && pytest -v test_python.py
+          cd clients/python && python3 umbridge-client.py http://model:4242
