### 2022-11-21 #4748 @linusseelinger
@Himscipy Thank you for your feedback! I have added some tests to fully cover the C++ and Python clients beyond the examples we had. The server side should be fairly well-tested already (see protocol_conformity_1.0, which is run on every model). I'm hesitant about code coverage checks, since UM-Bridge includes a couple of different languages, and more are to come. Also, since the libraries are mostly protocol wrappers, full coverage is fairly easy to achieve.
Will soon adapt the paper as requested!
