### 2022-11-23 39f71bc
Streamline paper, add comparison to other packages
