+++ joss/paper.bib
@@ -106,3 +106,23 @@ pages = {},
 title = {Probabilistic programming in Python using PyMC3},
 doi = {10.7287/PEERJ.PREPRINTS.1686V1}
 }
+
+@Article{preCICE,
+AUTHOR = { Chourdakis, G and Davis, K and Rodenberg, B and Schulte, M and Simonis, F and Uekermann, B and Abrams, G and Bungartz, HJ and Cheung Yau, L and Desai, I and Eder, K and Hertrich, R and Lindner, F and Rusch, A and Sashko, D and Schneider, D and Totounferoush, A and Volland, D and Vollmer, P and Koseomur, OZ},
+TITLE = {preCICE v2: A sustainable and user-friendly coupling library [version 2; peer review: 2 approved]
+},
+JOURNAL = {Open Research Europe},
+VOLUME = {2},
+YEAR = {2022},
+NUMBER = {51},
+DOI = {10.12688/openreseurope.14445.2}
+}
+
+@misc{QMCPy,
+  Author = {S.-C. T. Choi and F. J. Hickernell and M. McCourt and A. Sorokin},
+  Date-Added = {2020-04-15 15:19:14 -0500},
+  Date-Modified = {2020-04-26 17:13:25 -0500},
+  Title = {{QMCPy}: A quasi-{M}onte {C}arlo {P}ython Library},
+  Url = {https://github.com/QMCSoftware/QMCSoftware},
+  Year = {2020+},
+  Bdsk-Url-1 = {https://github.com/QMCSoftware/QMCSoftware}}
