+++ joss/paper.md
@@ -4,6 +4,7 @@ title: 'UM-Bridge: Uncertainty quantification and modeling bridge'
 tags:
   - Python
   - C++
+  - R
 authors:
   - name: Linus Seelinger
     orcid: 0000-0001-8632-8493
@@ -30,9 +31,9 @@ affiliations:
    index: 1
  - name: Courant Institute of Mathematical Sciences, New York Univeristy, New York, NY, USA
    index: 2
- - name: Department of Computer Science, Durham University, Durham, United Kingdom 
+ - name: Department of Computer Science, Durham University, Durham, United Kingdom
    index: 3
- - name: Department of Mathematics, Dartmouth College, Hannover, NH, USA 
+ - name: Department of Mathematics, Dartmouth College, Hannover, NH, USA
    index: 4
  - name: Independent researcher
    index: 5
@@ -43,29 +44,32 @@ bibliography: paper.bib
 
 # Summary
 
-UM-Bridge (the uncertainty quantification (UQ) and Modeling Bridge) provides a unified interface for numerical models that is accessible from any programming language or framework. It is primarily intended for coupling advanced models (e.g. simulations of complex physical processes) to advanced statistical or optimization methods for UQ without requiring the model and UQ algorithms to share a common computational environment.
+UM-Bridge (the uncertainty quantification (UQ) and modeling bridge) provides a unified interface for numerical models that is accessible from any programming language or framework. It is primarily intended for coupling advanced models (e.g. simulations of complex physical processes) to advanced statistical or optimization methods for UQ without requiring the model and UQ algorithms to share a common computational environment.
 
-Many uncertainty quantification and optimization methods treat a model as an abstract function $f: \mathbb{R}^{n} \mapsto \mathbb{R}^{m}$ and only interact with the model through some of the following operations: (i) model evaluation $y = f(x)$, (ii) gradient evaluation, (iii) Jacobian action, and/or (iv) Hessian action. The UQ algorithms rarely require detailed knowledge of $f$ other than these abstracted functions. In theory, this abstraction allows the same implementation of a UQ algorithm to be easily applied on a wide range of application problems. In practice however, the challenge of building and using UQ software and models in the same environment can limit, or even prevent entirely, the use of UQ by model developers. This issue is exacerbated by the distributed or heterogeneous computing environments used required by many advanced simulation codes; ensuring both UQ and simulation codes are built and interoperable in such an environment is a significant barrier.
+By allowing containerization, numerical models become portable and reproducible. This improves separation of concerns between the fields, and lays the groundwork for unified UQ benchmark problems. Further, high peformance computing is simplified through a pre-defined configuration for cloud environments, allowing to run many parallel instances of any UM-Bridge model container.
 
 # Statement of need
 
-UM-Bridge implements a software interface that mirrors the mathematical "interface" between models and UQ algorithms. Many UQ algorithms use a "black-box" model $f$, often a physics-based system of equations or a statistical model. Examples include Markov chain Monte Carlo methods [@MHMCMC; @ParallelMLMCMC], polynomial chaos [@PCE], stochastic collocation [@Marzouk_StochasticCollocation], optimal transport [@SamplingTransportMaps], and maximum likelihood estimation. UQ algorithms and models are often developed separately. Each implementation is typically done by experts in different fields that do not necessarily design compatible software. Implementing interfaces between these code bases is time consuming, tedious, and sometimes requires completely re-implementing either the UQ algorithm or the model. UM-Bridge addresses this challenge by providing a software framework that allows any model to be practically treated like a black-box.
[…]
