### 2022-11-23 09674bc
Remove command override from MPI launcher yml
