### 2022-11-23 f89edfe
Note GCE disk in kubernetes setup
