### 2022-11-28 fea4373
Switch QMCPy test to QMCPy from pip
