+++ .github/workflows/client-python-tinyDA.yml
@@ -1,0 +1,31 @@
+name: client-python-tinyDA
+
+on:
+  push:
+    branches:
+      - 'main'
+
+jobs:
+
+  test:
+    runs-on: ubuntu-latest
+    container: ubuntu:latest
+
+    services:
+      model:
+        image: linusseelinger/model-muq-beam:latest
+        ports:
+        - 4242:4242
+
+    steps:
+       -
+        name: Checkout
+        uses: actions/checkout@v2
+       -
+        name: Dependencies
+        run: |
+          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install papermill umbridge tinyDA
+       -
+        name: Build and run
+        run: |
+          papermill -k python3 --progress-bar clients/python/tinyDA-client.ipynb - > /dev/null
