+++ .github/workflows/client-python-tinyDA.yml
@@ -24,7 +24,7 @@ jobs:
        -
         name: Dependencies
         run: |
-          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install papermill umbridge tinyDA
+          apt update && DEBIAN_FRONTEND="noninteractive" apt install -y python3-pip && pip3 install papermill umbridge tinyDA ipykernel
        -
         name: Build and run
         run: |
