### 2022-12-10 d339ace
Fix k8s docs for getting ingress IP
