### 2022-12-12 a3d7ce1
class interface, constructor, get input size
