### 2022-12-12 3e12760
get_output_sizes in Matlab client
