+++ clients/matlab/@HTTPModel/get_output_sizes.m
@@ -1,4 +1,7 @@
-function output_sizes = get_output_sizes(self)
+function output_sizes = get_output_sizes(self, config)
+if (nargin<2) || (isempty(config))
+    config = struct;  % will result in {} in json. Order of fields does not matter
+end
 
 % this should only import the functions that we need
 import matlab.net.*
@@ -6,7 +9,7 @@ import matlab.net.http.*
 
 r = RequestMessage('POST');
 value.name = self.model_name;
-value.config=struct; % will result in {} in json. Order of fields does not matter
+value.config = config;
 r.Body = MessageBody(jsonencode(value));
 %disp(r.Body.string);
 uri = URI([self.uri,'/OutputSizes']);
