### 2022-12-13 f080510
check that parfor works + protocol + polishing error control
