### 2022-12-14 f36bf6a
adds a install config and installed config
