+++ testing/clients/test_matlab.m
@@ -1,0 +1,16 @@
+classdef TestUmbridge < matlab.unittest.TestCase
+
+    methods (Test)
+
+        function testConnection(testCase)
+            uri = 'http://model:4243';
+
+            model = HTTPModel(uri,'posterior');
+
+            httpValue = model.evaluate([1, 3])
+            exactValue = -11.194036030183454;
+            testCase.verifyEqual(httpValue, exactValue, 'RelTol', 1e-14)
+
+        end
+    end
+end
