+++ .github/workflows/client-matlab.yml
@@ -9,6 +9,7 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
+    container: ubuntu:latest
 
     services:
       model:
