+++ .github/workflows/client-matlab.yml
@@ -9,7 +9,6 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
-    container: ubuntu:latest
 
     services:
       model:
