+++ .github/workflows/client-matlab.yml
@@ -25,6 +25,7 @@ jobs:
       - name: Run all tests
         uses: matlab-actions/run-tests@v1
         with:
+          source-folder: matlab
           source-folder: clients/matlab
 
       #As an alternative to run-tests, you can use run-command to execute a MATLAB script, function, or statement.
