### 2023-02-22 805b9ad
Improved handling of connections in Python
