### 2023-03-15 #4748 @pdebuyl
Hi @linusseelinger I read the paper and tested a few of the examples. Everything ran flawlessly and I have no further comments or recommendations.

Can you
1. Make a tagged release
2. archive it on zenodo (make sure to use same title and authors as JOSS article)
3. provide the doi of the zenodo archive here
