### 2023-03-17 #4748 @danielskatz
👋 @linusseelinger - I'm the track editor who will take over at this point. While proofreading the paper, I found some small changes that I think are needed: please merge https://github.com/UM-Bridge/umbridge/pull/12 or let me know what you disagree with, then we can proceed to acceptance and publication
